+++ finance/portfolios/portfolios.py
@@ -75,17 +75,21 @@ class PortfolioStore(object):
         for s in self.holdings:
             t = self.holdings[s]["object"].parent_store.type
             if t == "company":
-                for h in self.holdings[s]["holding_measures"]:
-                    total_mkt_value += h["Base Mkt Val"]
-                    initial_weight += h["Portfolio_Weight"]
+                total_mkt_value += self.holdings[s]["holding_measures"]["Base Mkt Val"]
+                initial_weight += self.holdings[s]["holding_measures"][
+                    "Portfolio_Weight"
+                ]
             elif t == "sovereign":
-                for h in self.holdings[s]["holding_measures"]:
-                    total_mkt_value_sov += h["Base Mkt Val"]
-                    initial_weight_sov += h["Portfolio_Weight"]
+                total_mkt_value_sov += self.holdings[s]["holding_measures"][
+                    "Base Mkt Val"
+                ]
+                initial_weight_sov += self.holdings[s]["holding_measures"][
+                    "Portfolio_Weight"
+                ]
 
-        self.total_market_value_corp = float(total_mkt_value) / exchange_rate
+        self.total_market_value_corp = total_mkt_value / exchange_rate
         self.initial_weight_corp = initial_weight
-        self.total_market_value_sov = float(total_mkt_value_sov) / exchange_rate
+        self.total_market_value_sov = total_mkt_value_sov / exchange_rate
         self.initial_weight_sov = initial_weight_sov
 
     def calculate_carbon_impact(self, impact_column: str) -> None:
@@ -158,27 +162,27 @@ class PortfolioStore(object):
             energy = self.holdings[s]["object"].parent_store.msci_information[
                 "ENERGY_CONSUMP_INTEN_EUR"
             ]
-            for h in self.holdings[s]["holding_measures"]:
-                weight = h["Portfolio_Weight"]
 
-                if (
-                    not (pd.isna(nace_code) or pd.isna(energy) or weight == 0)
-                    and t == "company"
-                    and nace_code == nace_section_code
-                ):
-                    total_weight += weight
-                    data.append(
-                        {
-                            "ISIN": s,
-                            "NACE_SECTION_CODE": nace_code,
-                            "ENERGY_CONSUMP_INTEN_EUR": energy,
-                            "Portfolio_Weight": weight,
-                        }
-                    )
+            weight = self.holdings[s]["holding_measures"]["Portfolio_Weight"]
+
+            if (
+                not (pd.isna(nace_code) or pd.isna(energy) or weight == 0)
+                and t == "company"
+                and nace_code == nace_section_code
+            ):
+                total_weight += weight
+                data.append(
+                    {
+                        "ISIN": s,
+                        "NACE_SECTION_CODE": nace_code,
+                        "ENERGY_CONSUMP_INTEN_EUR": energy,
+                        "Portfolio_Weight": weight,
+                    }
+                )
 
         impact = 0
         for s in data:
-            norm_weight = float(s["Portfolio_Weight"] / total_weight)
+            norm_weight = s["Portfolio_Weight"] / total_weight
             impact += s["ENERGY_CONSUMP_INTEN_EUR"] * norm_weight
         coverage = (
             total_weight / self.initial_weight_corp
@@ -257,22 +261,22 @@ class PortfolioStore(object):
             value = self.holdings[s]["object"].parent_store.msci_information[
                 "CTRY_GHG_INTEN_GDP_EUR"
             ]
-            for h in self.holdings[s]["holding_measures"]:
-                weight = h["Portfolio_Weight"]
 
-                if not (pd.isna(value) or weight == 0) and t == "sovereign":
-                    total_weight += weight
-                    data.append(
-                        {
-                            "ISIN": s,
-                            "CTRY_GHG_INTEN_GDP_EUR": value,
-                            "Portfolio_Weight": weight,
-                        }
-                    )
+            weight = self.holdings[s]["holding_measures"]["Portfolio_Weight"]
+
+            if not (pd.isna(value) or weight == 0) and t == "sovereign":
+                total_weight += weight
+                data.append(
+                    {
+                        "ISIN": s,
+                        "CTRY_GHG_INTEN_GDP_EUR": value,
+                        "Portfolio_Weight": weight,
+                    }
+                )
 
         impact = 0
         for s in data:
-            norm_weight = float(s["Portfolio_Weight"] / total_weight)
+            norm_weight = s["Portfolio_Weight"] / total_weight
             impact += s["CTRY_GHG_INTEN_GDP_EUR"] * norm_weight
         coverage = (
             total_weight / self.initial_weight_sov
@@ -300,26 +304,26 @@ class PortfolioStore(object):
             value2 = self.holdings[s]["object"].parent_store.msci_information[
                 "GOVERNMENT_UN_SANCTIONS"
             ]
-            for h in self.holdings[s]["holding_measures"]:
-                weight = h["Portfolio_Weight"]
 
-                if (
-                    not (pd.isna(value1) or pd.isna(value2) or weight == 0)
-                    and t == "sovereign"
-                ):
-                    total_weight += weight
-                    data.append(
-                        {
-                            "ISIN": s,
-                            "GOVERNMENT_EU_SANCTIONS": value1,
-                            "GOVERNMENT_UN_SANCTIONS": value2,
-                            "Portfolio_Weight": weight,
-                        }
-                    )
+            weight = self.holdings[s]["holding_measures"]["Portfolio_Weight"]
+
+            if (
+                not (pd.isna(value1) or pd.isna(value2) or weight == 0)
+                and t == "sovereign"
+            ):
+                total_weight += weight
+                data.append(
+                    {
+                        "ISIN": s,
+                        "GOVERNMENT_EU_SANCTIONS": value1,
+                        "GOVERNMENT_UN_SANCTIONS": value2,
+                        "Portfolio_Weight": weight,
+                    }
+                )
 
         impact = 0
         for s in data:
-            norm_weight = float(s["Portfolio_Weight"] / total_weight)
+            norm_weight = s["Portfolio_Weight"] / total_weight
             if (
                 s["GOVERNMENT_EU_SANCTIONS"] == "Yes"
                 or s["GOVERNMENT_UN_SANCTIONS"] == "Yes"
@@ -390,26 +394,23 @@ class PortfolioStore(object):
             value = self.holdings[s]["object"].parent_store.msci_information[
                 impact_column
             ]
-            for h in self.holdings[s]["holding_measures"]:
-                weight = h["Portfolio_Weight"]
 
-                if (
-                    not (pd.isna(evic) or pd.isna(value) or weight == 0)
-                    and t == "company"
-                ):
-                    total_weight += weight
-                    data.append(
-                        {
-                            "ISIN": s,
-                            "EVIC_EUR": evic,
-                            impact_column: value,
-                            "Portfolio_Weight": weight,
-                        }
-                    )
+            weight = self.holdings[s]["holding_measures"]["Portfolio_Weight"]
+
+            if not (pd.isna(evic) or pd.isna(value) or weight == 0) and t == "company":
+                total_weight += weight
+                data.append(
+                    {
+                        "ISIN": s,
+                        "EVIC_EUR": evic,
+                        impact_column: value,
+                        "Portfolio_Weight": weight,
+                    }
+                )
 
         impact = 0
         for s in data:
-            norm_weight = float(s["Portfolio_Weight"] / total_weight)
+            norm_weight = s["Portfolio_Weight"] / total_weight
             investor_stake = (
                 norm_weight * self.total_market_value_corp / (s["EVIC_EUR"] * 1000000)
             )
@@ -452,22 +453,22 @@ class PortfolioStore(object):
             value = self.holdings[s]["object"].parent_store.msci_information[
                 impact_column
             ]
-            for h in self.holdings[s]["holding_measures"]:
-                weight = h["Portfolio_Weight"]
 
-                if not (pd.isna(value) or weight == 0) and t == "company":
-                    total_weight += weight
-                    data.append(
-                        {
-                            "ISIN": s,
-                            impact_column: value,
-                            "Portfolio_Weight": weight,
-                        }
-                    )
+            weight = self.holdings[s]["holding_measures"]["Portfolio_Weight"]
+
+            if not (pd.isna(value) or weight == 0) and t == "company":
+                total_weight += weight
+                data.append(
+                    {
+                        "ISIN": s,
+                        impact_column: value,
+                        "Portfolio_Weight": weight,
+                    }
+                )
 
         impact = 0
         for s in data:
-            norm_weight = float(s["Portfolio_Weight"] / total_weight)
+            norm_weight = s["Portfolio_Weight"] / total_weight
             waci = norm_weight * s[impact_column]
             impact += waci
 
@@ -505,22 +506,22 @@ class PortfolioStore(object):
             value = self.holdings[s]["object"].parent_store.msci_information[
                 impact_column
             ]
-            for h in self.holdings[s]["holding_measures"]:
-                weight = h["Portfolio_Weight"]
 
-                if not (pd.isna(value) or weight == 0) and t == "company":
-                    total_weight += weight
-                    data.append(
-                        {
-                            "ISIN": s,
-                            impact_column: value,
-                            "Portfolio_Weight": weight,
-                        }
-                    )
+            weight = self.holdings[s]["holding_measures"]["Portfolio_Weight"]
+
+            if not (pd.isna(value) or weight == 0) and t == "company":
+                total_weight += weight
+                data.append(
+                    {
+                        "ISIN": s,
+                        impact_column: value,
+                        "Portfolio_Weight": weight,
+                    }
+                )
 
         impact = 0
         for s in data:
-            norm_weight = float(s["Portfolio_Weight"] / total_weight)
+            norm_weight = s["Portfolio_Weight"] / total_weight
             if s[impact_column] == filter_word:
                 impact += norm_weight
         coverage = (